refactor(srbench): rename conversation.py to signals.py, move action counting to executors
ConversationSignals now owns only coordination (wake events, end event,
stalemate detection). The action count was telemetry riding along on the
shared object, so it moves to its consumers. The calendar executor keeps
a conversation-wide counter shared by both invoke callbacks, and the
marketplace executor derives total_actions from len(action_trace) since
every invoke appends exactly one trace entry.
